chore: reorganize hotkey and config
