+++ BA-Cheeto.vcxproj
@@ -16,8 +16,11 @@
     <ClInclude Include="src\appdata\cs-types.h" />
     <ClInclude Include="src\appdata\types-helper.h" />
     <ClInclude Include="src\appdata\types.h" />
+    <ClInclude Include="src\core\config\config_field.h" />
+    <ClInclude Include="src\core\config\config_manager.h" />
     <ClInclude Include="src\core\events\event.h" />
     <ClInclude Include="src\core\events\event_manager.h" />
+    <ClInclude Include="src\core\hotkey\hotkey_manager.h" />
     <ClInclude Include="src\core\pipe\pipe_manager.h" />
     <ClInclude Include="src\core\rendering\backend\dx11_backend.h" />
     <ClInclude Include="src\core\rendering\backend\renderer_backend.h" />
@@ -41,10 +44,7 @@
     <ClInclude Include="src\user\cheat\feature_base.h" />
     <ClInclude Include="src\user\cheat\feature_manager.h" />
     <ClInclude Include="src\user\main.h" />
-    <ClInclude Include="src\utils\config_field.h" />
-    <ClInclude Include="src\utils\config_manager.h" />
     <ClInclude Include="src\utils\dx_utils.h" />
-    <ClInclude Include="src\utils\hotkey_manager.h" />
     <ClInclude Include="src\utils\logger.h" />
     <ClInclude Include="src\utils\singleton.h" />
     <ClInclude Include="vendor\imgui\backends\imgui_impl_dx11.h" />
@@ -75,7 +75,9 @@
     <ClInclude Include="vendor\UnityResolve\UnityResolve.hpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="src\core\config\config_manager.cpp" />
     <ClCompile Include="src\core\events\event_manager.cpp" />
+    <ClCompile Include="src\core\hotkey\hotkey_manager.cpp" />
     <ClCompile Include="src\core\pipe\pipe_manager.cpp" />
     <ClCompile Include="src\core\rendering\backend\dx11_backend.cpp" />
     <ClCompile Include="src\core\rendering\renderer.cpp" />
@@ -99,9 +101,7 @@
     <ClCompile Include="src\user\cheat\features\player\PlayerStats.cpp" />
     <ClCompile Include="src\user\cheat\feature_manager.cpp" />
     <ClCompile Include="src\user\main.cpp" />
-    <ClCompile Include="src\utils\config_manager.cpp" />
     <ClCompile Include="src\utils\dx_utils.cpp" />
-    <ClCompile Include="src\utils\hotkey_manager.cpp" />
     <ClCompile Include="src\utils\logger.cpp" />
     <ClCompile Include="vendor\imgui\backends\imgui_impl_dx11.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
