+++ src/core/config/config_manager.h
