+++ src/core/hotkey/hotkey_manager.h
