+++ src/dllmain.cpp
@@ -1,6 +1,6 @@
 #include "pch.h"
 #include "user/main.h"
-#include "utils/config_manager.h"
+#include "core/config/config_manager.h"
 #include <filesystem>
 #include <thread>
